pyproject.toml
[project]
name = "property-manage-database"
version = ""
description = ""
authors = [
    {name = "xxc", email = "x007007007@hotmail.com"},
]
packages = [
    "src"
]

dependencies = [
    "django~=3.2",
    "djangorestframework~=3.12",
    "markdown~=3.3",
    "django-filter~=21.1",
    "django-guardian~=2.4",
    "Pygments~=2.10",
    "drf-yasg[validation]~=1.20",
    "graphene-django~=2.15",
    "python-dateutil~=2.8",
    "django-debug-toolbar~=3.2",
    "django-cors-headers~=3.10",
    "django-admin-rangefilter~=0.8",
    "django-admin-list-filter-dropdown~=1.0",
    "django-extensions~=3.1",
    "aiortc~=1.2",
    "django-admin-charts~=0.24",
    "channels~=3.0",
    "channels-redis~=3.3",
]
requires-python = ">=3.7"
license = {text = "MIT"}

[project.urls]
homepage = ""

[build-system]
requires = ["pdm-pep517"]
build-backend = "pdm.pep517.api"

[tool]
[tool.pdm]
